refactor: relocate useToast
