+++ src/app/(private)/admin-registration-applications/[id]/page.tsx
@@ -12,8 +12,8 @@ import {
 } from '@/components/ui/form'
 import { Input } from '@/components/ui/input'
 import { LoadingSpinner } from '@/components/ui/loadingSpinner'
-import { useToast } from '@/components/ui/use-toast'
 import { Permission } from '@/constants/permissions'
+import { useToast } from '@/hooks/useToast'
 import { formatDateTime, formatPhoneNumber } from '@/lib/dataFormatters'
 import { handleApiError } from '@/lib/handleApiError'
 import { FormValidationSchema } from '@/modules/adminRegistrationApplications/constants/formValidationSchema'
