+++ src/app/(private)/emergency-evacuation-applications/page.tsx
@@ -8,13 +8,13 @@ import FilterInput from '@/components/ui/filter-input'
 import MultiSelectDropdown from '@/components/ui/multi-select-dropdown'
 import Status from '@/components/ui/status'
 import { Toaster } from '@/components/ui/toaster'
-import { toast } from '@/components/ui/use-toast'
 import { getStringFilterValidation } from '@/constants/filterValidationSchema'
 import { Permission } from '@/constants/permissions'
 import useDebouncedInputFilter from '@/hooks/useDebouncedInputFilter'
 import { useHandleFilterChange } from '@/hooks/useHandleFilterChange'
 import { usePagination } from '@/hooks/usePagination'
 import { useSort } from '@/hooks/useSort'
+import { toast } from '@/hooks/useToast'
 import { handleApiError } from '@/lib/handleApiError'
 import { selectPermissions } from '@/modules/auth/authSlice'
 import { columns } from '@/modules/emergencyEvacuationApplications/components/columns'
