+++ src/app/(private)/users/page.tsx
@@ -1,31 +1,31 @@
 'use client'
 
+import { FilterValidationOptions, Sort } from '@/common/types'
 import { Button } from '@/components/ui/button'
 import { DataTable } from '@/components/ui/data-table'
-import { Toaster } from '@/components/ui/toaster'
-import { usePagination } from '@/hooks/usePagination'
-import { handleApiError } from '@/lib/handleApiError'
-import { useRouter, useSearchParams, usePathname } from 'next/navigation'
-import { useCallback, useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
-import { useSort } from '@/hooks/useSort'
-import { useHandleFilterChange } from '@/hooks/useHandleFilterChange'
+import FilterInput from '@/components/ui/filter-input'
 import MultiSelectDropdown from '@/components/ui/multi-select-dropdown'
 import Status from '@/components/ui/status'
+import { Toaster } from '@/components/ui/toaster'
+import { getStringFilterValidation } from '@/constants/filterValidationSchema'
+import { Permission } from '@/constants/permissions'
+import useDebouncedInputFilter from '@/hooks/useDebouncedInputFilter'
+import { useHandleFilterChange } from '@/hooks/useHandleFilterChange'
+import { usePagination } from '@/hooks/usePagination'
+import { useSort } from '@/hooks/useSort'
+import { toast } from '@/hooks/useToast'
+import { handleApiError } from '@/lib/handleApiError'
+import { selectPermissions } from '@/modules/auth/authSlice'
 import { columns, User } from '@/modules/users/components/columns'
+import { userStatuses } from '@/modules/users/constants/statuses'
 import { UsersFilter } from '@/modules/users/constants/types'
 import { getUsers } from '@/modules/users/service'
-import FilterInput from '@/components/ui/filter-input'
-import { RefreshCw } from 'lucide-react'
-import { userStatuses } from '@/modules/users/constants/statuses'
-import { FilterValidationOptions, Sort } from '@/common/types'
-import { getStringFilterValidation } from '@/constants/filterValidationSchema'
-import { toast } from '@/components/ui/use-toast'
 import { useAppSelector } from '@/store/hooks'
-import { selectPermissions } from '@/modules/auth/authSlice'
-import { Permission } from '@/constants/permissions'
+import { RefreshCw } from 'lucide-react'
 import Link from 'next/link'
-import useDebouncedInputFilter from '@/hooks/useDebouncedInputFilter'
+import { usePathname, useRouter, useSearchParams } from 'next/navigation'
+import { useCallback, useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
 
 const parseUsersSearchParams = (searchParams: URLSearchParams) => {
   const currentPage = parseInt(searchParams.get('page') ?? '1', 10)
